Update c02z09v1.cpp
I have changed the algorithm a bit. You don't have to increase sum2 till infinity. All you have to do is to check for the smallest value of sum2 that is bigger or equal to n
I have left the comments, in case you wanted to know how it works.
Sorry for my English xDD